chore(ErrorDetails): move to containers [YTFRONT-5665]
