+++ packages/ui/src/ui/containers/ErrorDetails/ErrorDetails.scss
