+++ packages/ui/src/ui/containers/ErrorDetails/i18n/dicts.ts
