+++ packages/ui/src/ui/containers/ErrorDetails/i18n/en.json
