+++ packages/ui/src/ui/pages/navigation/Navigation/NavigationError/NavigationError.tsx
@@ -3,7 +3,7 @@ import cn from 'bem-cn-lite';
 import React from 'react';
 import {type YTError} from '../../../../../@types/types';
 import {YTErrorBlock} from '../../../../containers/Block/Block';
-import ErrorDetails from '../../../../components/ErrorDetails/ErrorDetails';
+import ErrorDetails from '../../../../containers/ErrorDetails/ErrorDetails';
 import {ErrorToClipboardButton} from '../../../../containers/ErrorToClipboardButton/ErrorToClipboardButton';
 import NavigationDescription from '../../../../pages/navigation/NavigationDescription/NavigationDescription';
 import {getPermissionDeniedError} from '../../../../utils/errors';
