+++ packages/ui/src/ui/pages/operations/OperationDetail/tabs/incarnations/IncarnationInfoDialog.scss
@@ -1,4 +1,4 @@
-@use '../../../../../components/ErrorDetails/ErrorDetails.scss' as error;
+@use '../../../../../containers/ErrorDetails/ErrorDetails.scss' as error;
 
 .incarnation-info-dialog {
     &__details {
